rename txn to tx
