+++ cmd/soroban-cli/src/commands/tx/sign.rs
