+++ cmd/soroban-cli/src/commands/tx/simulate.rs
